requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
faiss-cpu
sentence-transformers
PyPDF2
python-docx
